Agregado sistemas P.U

+++ structs.jl
@@ -1,9 +1,9 @@
 struct Generador
     ID :: Vector{Int64}
-    Pmin :: Vector{Int64}
+    Pmin :: Vector{Float64}
     Pmax :: Vector{Float64}
-    Cvariable :: Vector{Int64}
-    Ramp :: Vector{Int64}
+    Cvariable :: Vector{Float64}
+    Ramp :: Vector{Float64}
     Bus_Conexion :: Vector{Int64}
 end
 
@@ -17,19 +17,19 @@ end
 
 struct Loads
     ID_Bus :: Vector{Int64}
-    D_t1 :: Vector{Int64}
-    D_t2 :: Vector{Int64}
-    D_t3 :: Vector{Int64}
-    D_t4 :: Vector{Int64}
-    D_t5 :: Vector{Int64}
-    D_t6 :: Vector{Int64}
+    D_t1 :: Vector{Float64}
+    D_t2 :: Vector{Float64}
+    D_t3 :: Vector{Float64}
+    D_t4 :: Vector{Float64}
+    D_t5 :: Vector{Float64}
+    D_t6 :: Vector{Float64}
 end
 
 struct Lines
     ID :: Vector{Int64}
     FromBus :: Vector{Int64}
     ToBus :: Vector{Int64}
-    Fmax :: Vector{Int64}
+    Fmax :: Vector{Float64}
     X :: Vector{Float64}
     
 end
